+++ flybywire-aircraft-a320-neo/html_ui/Pages/VCockpit/Instruments/Airliners/FlyByWire_A320_Neo/CDU/A320_Neo_CDU_MainDisplay.js
@@ -19,6 +19,7 @@ class A320_Neo_CDU_MainDisplay extends FMCMainDisplay {
         this.sentMessages = [];
         this.activeSystem = 'FMGC';
         this.messageQueue = [];
+        this.aocAirportList = undefined;
     }
     get templateID() {
         return "A320_Neo_CDU";
@@ -36,10 +37,12 @@ class A320_Neo_CDU_MainDisplay extends FMCMainDisplay {
 
     initMcduVariables() {
         this.messageQueue = [];
+        this.aocAirportList = new CDUAocAirportList;
     }
 
     Init() {
         super.Init();
+        this.initMcduVariables();
 
         let mainFrame = this.getChildById("Mainframe");
         if (mainFrame == null) {