+++ flybywire-aircraft-a320-neo/html_ui/Pages/VCockpit/Instruments/Airliners/FlyByWire_A320_Neo/CDU/ATSU/A320_Neo_CDU_AOC_RequestsWeather.js
@@ -1,14 +1,15 @@
 class CDUAocRequestsWeather {
-    static ShowPage(mcdu, setArpts = true, store = {"reqID": 0, "depIcao": "", "arrIcao": "", "arpt1": "", "arpt2": "", "arpt3": "", "arpt4": "", "sendStatus": ""}) {
+    static ShowPage(mcdu, reqID = 0, _sendStatus = "") {
         mcdu.clearDisplay();
         let labelTimeout;
+        let sendStatus = _sendStatus;
 
-        if (mcdu.flightPlanManager.getOrigin() && mcdu.flightPlanManager.getDestination()) {
-            if (setArpts) {
-                store.arpt1 = mcdu.flightPlanManager.getOrigin().ident;
-                store.arpt2 = mcdu.flightPlanManager.getDestination().ident;
-            }
-        }
+        // if (mcdu.flightPlanManager.getOrigin() && mcdu.flightPlanManager.getDestination()) {
+        //     if (setArpts) {
+        //         store.arpt1 = mcdu.flightPlanManager.getOrigin().ident;
+        //         store.arpt2 = mcdu.flightPlanManager.getDestination().ident;
+        //     }
+        // }
 
         const reqTypes = [
             'METAR',
@@ -19,71 +20,39 @@ class CDUAocRequestsWeather {
             mcdu.setTemplate([
                 ["AOC WEATHER REQUEST"],
                 [`WX TYPE`, "AIRPORTS"],
-                [`↓${reqTypes[store.reqID]}[color]cyan`, `${store.arpt1 != "" ? store.arpt1 : "[ ]"}[color]cyan`],
+                [`↓${reqTypes[reqID]}[color]cyan`, mcdu.aocAirportList.rows[0].outPut],
                 [""],
-                ["", `${store["arpt2"] != "" ? store["arpt2"] : "[ ]"}[color]cyan`],
+                ["", mcdu.aocAirportList.rows[1].outPut],
                 [""],
-                ["", `${store["arpt3"] != "" ? store["arpt3"] : "[ ]"}[color]cyan`],
+                ["", mcdu.aocAirportList.rows[2].outPut],
                 [""],
-                ["", `${store["arpt4"] != "" ? store["arpt4"] : "[ ]"}[color]cyan`],
+                ["", mcdu.aocAirportList.rows[3].outPut],
                 [""],
                 [""],
-                ["RETURN TO", `${store["sendStatus"]}`],
+                ["RETURN TO", `${sendStatus}`],
                 ["<AOC MENU", "SEND*[color]cyan"]
             ]);
         };
         updateView();
 
-        mcdu.rightInputDelay[0] = () => {
-            return mcdu.getDelaySwitchPage();
-        };
-
         mcdu.onRightInput[0] = (value) => {
-            if (value === FMCMainDisplay.clrValue) {
-                store["arpt1"] = "";
-            } else {
-                store["arpt1"] = value;
-            }
-            CDUAocRequestsWeather.ShowPage(mcdu, false, store);
-        };
-
-        mcdu.rightInputDelay[1] = () => {
-            return mcdu.getDelaySwitchPage();
+            mcdu.aocAirportList.set(0, value);
+            CDUAocRequestsWeather.ShowPage(mcdu, reqID, sendStatus);
         };
 
         mcdu.onRightInput[1] = (value) => {
-            if (value === FMCMainDisplay.clrValue) {
-                store["arpt2"] = "";
-            } else {
-                store["arpt2"] = value;
-            }
-            CDUAocRequestsWeather.ShowPage(mcdu, false, store);
-        };
-
-        mcdu.rightInputDelay[2] = () => {
-            return mcdu.getDelaySwitchPage();
+            mcdu.aocAirportList.set(1, value);
+            CDUAocRequestsWeather.ShowPage(mcdu, reqID, sendStatus);
         };
 
         mcdu.onRightInput[2] = (value) => {
-            if (value === FMCMainDisplay.clrValue) {
-                store["arpt3"] = "";
-            } else {
-                store["arpt3"] = value;
-            }
-            CDUAocRequestsWeather.ShowPage(mcdu, false, store);
-        };
-
-        mcdu.rightInputDelay[3] = () => {
-            return mcdu.getDelaySwitchPage();
+            mcdu.aocAirportList.set(2, value);
+            CDUAocRequestsWeather.ShowPage(mcdu, reqID, sendStatus);
         };
 
         mcdu.onRightInput[3] = (value) => {
-            if (value === FMCMainDisplay.clrValue) {
-                store["arpt4"] = "";
-            } else {
-                store["arpt4"] = value;
-            }
-            CDUAocRequestsWeather.ShowPage(mcdu, false, store);
+            mcdu.aocAirportList.set(3, value);
+            CDUAocRequestsWeather.ShowPage(mcdu, reqID, sendStatus);
         };
 
         mcdu.rightInputDelay[5] = () => {
@@ -91,33 +60,33 @@ class CDUAocRequestsWeather {
         };
 
         mcdu.onRightInput[5] = async () => {
-            const icaos = [store["arpt1"], store["arpt2"], store["arpt3"], store["arpt4"]];
-            if (icaos.every((current) => current === "")) {
+            const icaos = mcdu.aocAirportList.icaos;
+            if (icaos.length === 0) {
                 mcdu.addNewMessage(NXFictionalMessages.noAirportSpecified);
                 return;
             }
-            store["sendStatus"] = "QUEUED";
+            sendStatus = "QUEUED";
             updateView();
             const lines = [];
-            const newMessage = { "id": Date.now(), "type": reqTypes[store.reqID], "time": '00:00', "opened": null, "content": lines, };
+            const newMessage = { "id": Date.now(), "type": reqTypes[reqID], "time": '00:00', "opened": null, "content": lines, };
             mcdu.clearUserInput();
 
             const getInfo = async () => {
-                if (store.reqID == 0) {
-                    getMETAR(icaos, lines, store, updateView);
+                if (reqID === 0) {
+                    getMETAR(icaos, lines, sendStatus, updateView);
                 } else {
-                    getTAF(icaos, lines, store, updateView);
+                    getTAF(icaos, lines, sendStatus, updateView);
                 }
             };
 
             getInfo().then(() => {
-                store["sendStatus"] = "SENT";
+                sendStatus = "SENT";
                 setTimeout(() => {
                     newMessage["time"] = fetchTimeValue();
                     mcdu.addMessage(newMessage);
                 }, Math.floor(Math.random() * 10000) + 10000);
                 labelTimeout = setTimeout(() => {
-                    store["sendStatus"] = "";
+                    sendStatus = "";
                     updateView();
                 }, 3000);
             });
@@ -127,8 +96,7 @@ class CDUAocRequestsWeather {
             return mcdu.getDelaySwitchPage();
         };
         mcdu.onLeftInput[0] = () => {
-            store["reqID"] = (store.reqID + 1) % 2;
-            CDUAocRequestsWeather.ShowPage(mcdu, false, store);
+            CDUAocRequestsWeather.ShowPage(mcdu, (reqID + 1) % 2, sendStatus);
         };
         mcdu.leftInputDelay[5] = () => {
             return mcdu.getDelaySwitchPage();