+++ .github/CHANGELOG.md
@@ -115,7 +115,7 @@
 1. [FMGC] Added reset of fmgc variables - @MisterChocker (Leon) 
 1. [ATHR] Fixed permanent A/THR disconnect (15s press) not being able to activate with 3D button on thrust levers - @aguther (Andreas Guther)
 1. [ND] TCAS symbols correctly display in 10, 20, 40nm ARC/ROSE views + bug fixes - @2hwk (2Cas#1022)
-1. [MCDU] Fixed airports on Weather Request page not editable - @SergeySoprykin (SergeyS) @MisterChocker (Leon)
+1. [MCDU] Reworked airport handling on AOC Weather Request Page - @MisterChocker (Leon)
 
 ## 0.6.0
 1. [CDU] Added WIND page - @tyler58546 (tyler58546)